+++ execution_engine_testing/tests/src/test/contract_api/mod.rs
@@ -16,3 +16,4 @@ mod subcall;
 mod transfer;
 mod transfer_purse_to_account;
 mod transfer_purse_to_purse;
+mod transfer_cached;