+++ execution_engine_testing/tests/src/test/contract_context.rs
@@ -364,8 +364,7 @@ fn should_not_call_session_from_contract() {
     builder.exec(exec_request_2).commit();
 
     let response = builder
-        .get_exec_results()
-        .last()
+        .get_last_exec_results()
         .expect("should have last response");
     assert_eq!(response.len(), 1);
     let exec_response = response.last().expect("should have response");