+++ execution_engine/CHANGELOG.md
@@ -13,6 +13,10 @@ All notable changes to this project will be documented in this file.  The format
 
 ## [Unreleased]
 
+### Changed
+
+* (Perf) Changed contract runtime to allow for executing many deploys per execute/commit cycle.
+
 
 ## [1.4.2] - 2021-11-11
 
@@ -28,8 +32,6 @@ All notable changes to this project will be documented in this file.  The format
 * Added genesis validation step to ensure there are more genesis validators than validator slots.
 * Added a support for passing a public key as a `target` argument in native transfers.
 * Added a `max_associated_keys` configuration option for a hard limit of associated keys under accounts.
-
-
 
 ### Changed
 * Documented `storage` module and children.